refactor: restructure Python modules
