+++ src/controllers/__init__.py
@@ -1,0 +1,1 @@
+from .theme_controller.theme_controller import ThemeController
