+++ src/controllers/theme_controller/__init__.py
@@ -1,0 +1,2 @@
+from .theme_controller_protocol import IThemeController
+from .theme_controller import ThemeController
