+++ src/ui/actions/__init__.py
@@ -1,0 +1,3 @@
+from .configure_connection_action import ConfigureConnectionAction
+from .quit_action import QuitAction
+from .tray_action_protocol import ITrayAction
