+++ src/ui/menu/__init__.py
@@ -1,0 +1,2 @@
+from .menu_factory_protocol import IMenuFactory
+from .menu_factory import MenuFactory
