+++ src/ui/system_tray/__init__.py
@@ -1,0 +1,2 @@
+from .system_tray_view_protocol import ISystemTrayIconView
+from .system_tray_view import SystemTrayIconView
