+++ src/ui/system_tray/system_tray.styles.qss
