+++ src/ui/system_tray/system_tray_view_protocol.py
